// ideqs/src/utils/enums.ts
/* eslint-disable no-unused-vars */
export enum SIDEBAR_ENUM {
  FORMATOS = 'FORMATOS',
  PUNTOS = 'PUNTOS',
  INFORMACION = 'INFORMACION',
  REMISIONES = 'REMISIONES',
  ORDERS_BUSINESS = 'ORDERS_BUSINESS',
  ORDERS_ADMINISTRATOR = 'ORDERS_ADMINISTRATOR',
  USUARIOS = 'USUARIOS',
  USUARIOSPRODUCCION = 'USUARIOSPRODUCCION',
  ORDERS_ENCARGADO = 'ORDERS_ENCARGADO',
  ORDERS_REMISION = 'ORDERS_REMISION'
}

export enum TYPE_ORDERS {
  PENDING = 1,
  ACCEPTED = 2,
  PRINTED = 3,
  FINISHED = 4,
  CANCELED = 5
}